strcpy

+++ 0x05-pointers_arrays_strings/9-strcpy.c
@@ -9,8 +9,8 @@
 
 char *_strcpy(char *dest, char *src)
 {
-	int i;
-//	dest = (char*)malloc(20);
+	long unsigned int i;
+
 	for (i =0; i < strlen(src); i++)
 	{
 		dest[i] = src [i];
